test package event type in changelog
